pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "py-cptcore"
version = "0.1.0"
description = "Public python SDK for the CEMS CPTCore web-API."
requires-python = ">=3.11"
dependencies = ["requests>=2.25.1,<3", "pandas>=2,<3", "matplotlib>=3,<4"]
license = { file = "LICENSE" }
readme = "README.md"
keywords = ["API", "CPTCore", "CEMS", "CRUX"]

[project.urls]
repository = "https://github.com/cemsbv/py-cptcore"

[project.optional-dependencies]
test = ["coveralls", "pytest"]
docs = [
  "Sphinx==8.2.3",
  "asteroid-sphinx-theme==0.0.3",
  "ipython==9.1.0",
  "sphinx-autodoc-typehints==3.1.0",
  "sphinx_rtd_theme==3.0.2",
]
gui = [
    "marimo[recommended]>0.14,<0.15",
    "micropip>0.10,<0.11",
    "cems-nuclei[client]>3,<4"
]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
files = ["pycptcore"]
mypy_path = 'src'
namespace_packages = true
show_error_codes = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true
disallow_subclassing_any = true
no_implicit_optional = true
disallow_untyped_defs = true
disallow_any_generics = false
disallow_untyped_calls = true
warn_return_any = false
no_implicit_reexport = true
strict_equality = true
install_types = true

[[tool.mypy.overrides]]
module = ["matplotlib.*", "requests.*", "pytest.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tests/*"]
disallow_untyped_defs = false
disallow_untyped_calls = false
